Fix external auth refresh for long queries.
When running queries that do not exceed the expiration of the access
token but are within the refresh grace period, make sure to remove
the "Authorization" header from CURL before reauthenticating. It's
still valid during the grace period, but we want to trigger a
re-auth intentionally.

Also fixed whitespace and PR workflow issues - growing pains from moving
the project to public github.