BUG: System DSNs decryptable by system users (#24)
One feature blocking the ODBC driver from working on the Power BI Gateway server was that System DSNs were usable by anyone on the system, but the client secret decryption was only possible by the user who created the DSN. This meant that when the Power BI Gateway Server tried to use the DSN via its service account, the client secret could not be decrypted.

With this commit in place, I have observed queries running via the Power BI Gateway server both for on-demand as well as scheduled data refreshes.